requirements_app.txt
amqp==5.2.0
async-timeout==4.0.3
bidict==0.22.1
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
h11==0.14.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
prompt-toolkit==3.0.42
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
redis==5.0.1
simple-websocket==1.0.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
zipp==3.17.0
